Improve package organization for PyPi release
