+++ permittivitycalc/__init__.py
@@ -1,0 +1,7 @@
+import numpy as np
+import uncertainties 
+#Citation: Uncertainties: a Python package for calculations with uncertainties,
+#    Eric O. LEBIGOT, http://pythonhosted.org/uncertainties/
+from uncertainties import unumpy as unp
+
+from .sparam_data import AirlineData, pplot, get_METAS_data, perm_compare, multiple_meas, run_default, run_example
